refactor: remove dead code and fold in one-liner fixes from code audit (Phase 3)
Delete confirmed-dead code with no call sites: the FlowLayout/FlowLayoutResult
custom Layout in TutorView, the segmentedPicker()/buttonLabel() view modifiers
and their backing structs in Modifiers, and Family.pastParticiplePrefix (a
second source of truth Conjugator never consulted).

Prune four serialized-but-unread WidgetSnapshot fields (frequency,
exampleEnglish, dateString, debugOffset) from the struct, WidgetSnapshotWriter,
and the SnapshotReader placeholder. The still-used debugOffset function
parameter (read from UserDefaults) and the static dateString(for:) helper (used
by questionID) are unaffected.

Fold in the audit one-liners: GameState.computeFinalScore Float to Double,
BrowseableFamily.topVerbs prefix(3).map { $0 } to Array(...prefix(3)), and
Quiz.timerInterval var to let (removing its redundant stored-property default
so init is the sole initializer).

Fix VerbParser diagnostics: the missing-infinitive message now reads the
parser's accumulating verbs.count instead of the still-empty global
Verb.verbs.count (was always reporting "position 1"), and add a duplicate-key
fatalError so a repeated in= attribute in Verbs.xml fails loudly instead of
silently overwriting an earlier verb.

Build succeeds and all 123 tests pass.

+++ Konjugieren/Models/GameState.swift
@@ -591,7 +591,7 @@ class GameState {
 
   func computeFinalScore() -> Int {
     let elapsed = Date.now.timeIntervalSince(startTime)
-    let raw = Float(score) * (Float(playerHealth) + 1.0) - Float(elapsed)
+    let raw = Double(score) * (Double(playerHealth) + 1.0) - Double(elapsed)
     return max(0, Int(raw))
   }
